Create working prototype
This implements a working proof of concept. The general algorithm
to determine the sinuosity is the following:
- Make bw
- Thin river
- Measure river length
- Measure distance between start and end

There are two assumptions in this algorithm:
1. The river must enter and exit the image some edge of the screen
and the only points of the river on the edge must be the start and
exit.
2. The contrast between the background and the river is
significant enough for thresholding the image to black and white
correctly.